pyproject.toml
[tool.poetry]
name = "keycode-parser"
version = "0.1.0"
description = ""
authors = ["ryzhovalex <thed4rkof@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "keycode_parser"}
]

[tool.poetry.dependencies]
python = "^3.11"
slimebones-pykit = "^0.4.14"
aiofiles = "^23.2.1"
slimebones-autofiles = "^0.1.1"

[tool.poetry.scripts]
keycode-parser = "keycode_parser.main:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.8"
coverage = "^7.3.4"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
